Add API methods

+++ backend/api.py
@@ -5,7 +5,7 @@ from flask_jwt_extended import current_user, jwt_required
 from peewee import PeeweeException, DoesNotExist
 from werkzeug.utils import secure_filename
 
-from .models import Category, Service, Subject, User, Contractor, Geography, ContractorUser
+from .models import db_wrapper, Category, Service, Subject, User, Contractor, Geography, ContractorUser
 
 
 bp = Blueprint('api', __name__)
@@ -14,17 +14,17 @@ bp = Blueprint('api', __name__)
 @bp.errorhandler(PeeweeException)
 def handle_data_error(e):
     current_app.logger.error(str(e))
-    return jsonify(error=str(e)), 400
+    return jsonify(msg=str(e)), 400
 
 
 @bp.errorhandler(DoesNotExist)
 def does_not_exist(e):
-    return jsonify(error=str(e)), 404
+    return jsonify(msg=str(e)), 404
 
 
 @bp.errorhandler(404)
 def resource_not_found(e):
-    return jsonify(error=str(e)), 404
+    return jsonify(msg=str(e)), 404
 
 
 @bp.route('/categories', methods=['GET'])
@@ -33,6 +33,41 @@ def list_categories():
         Category.select()
     )
     return [c.serialize for c in query]
+
+
+@bp.route('/categories', methods=['POST'])
+@jwt_required()
+def create_category():
+    if not current_user.admin:
+        return jsonify(msg='Доступ запрещён'), 401
+
+    data = request.get_json()
+    category = Category.create(**data)
+    return category.serialize
+
+
+@bp.route('/categories/<id>', methods=['PUT'])
+@jwt_required()
+def update_category(id):
+    if not current_user.admin:
+        return jsonify(msg='Доступ запрещён'), 401
+
+    data = request.get_json()
+    Category.update(**data).where(Category.id == id).execute()
+    category = Category.get_by_id(id)
+    return category.serialize
+
+
+@bp.route('/categories/<id>', methods=['DELETE'])
+@jwt_required()
+def delete_category(id):
+    if not current_user.admin:
+        return jsonify(msg='Доступ запрещён'), 401
+
+    with db_wrapper.database.atomic():
+        Service.delete().where(Service.category == id).execute()
+        Category.delete_by_id(id)
+    return id
 
 
 @bp.route('/services', methods=['GET'])
@@ -105,18 +140,21 @@ def list_contractors():
 
     if request.method == 'POST':
         data = request.get_json()
-        geography = data.pop('geography', [])
+        current_app.logger.error(data)
+        geography = data.pop('geography', None)
         if contractors.count() == 0:
             contractor = Contractor.create(**data)
             ContractorUser.create(contractor=contractor, user=current_user)
         else:
             contractor = contractors.first()
-            contractor.update(**data).execute()
-            contractor.reload()
-        Geography.delete().where(Geography.contractor == contractor).execute()
-        for code in geography:
-            subject = Subject.get(Subject.code == code)
-            Geography.create(contractor=contractor, subject=subject)
+            if data.keys():
+                contractor.update(**data).where(Contractor.id == contractor.id).execute()
+                contractor.reload()
+        if geography is not None:
+            Geography.delete().where(Geography.contractor == contractor).execute()
+            for code in geography:
+                subject = Subject.get(Subject.code == code)
+                Geography.create(contractor=contractor, subject=subject)
     elif contractors is None:
         return []
     return [contractor.serialize for contractor in contractors]
@@ -141,6 +179,20 @@ def get_contractor(id):
     #     contractor = contractors.where(User.id == current_user.id)
 
     return contractor.serialize
+
+
+@bp.route('/user/contractors/<id>', methods=['DELETE'])
+@jwt_required()
+def delete_contractor(id):
+    if not current_user.admin:
+        return jsonify(msg='Доступ запрещён'), 401
+
+    with db_wrapper.database.atomic():
+        Geography.delete().where(Geography.contractor == id).execute()
+        ContractorUser.delete().where(ContractorUser.contractor == id).execute()
+        Contractor.delete_by_id(id)
+
+    return id
 
 
 @bp.route('/user/files', methods=['POST'])
